% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGexpressionMapper.R
\name{KEGGexpressionMapper}
\alias{KEGGexpressionMapper}
\title{KEGGexpressionMapper}
\usage{
KEGGexpressionMapper(gene_desc = "example/slim_version/DEG.txt",
  gene_kegg = "example/slim_version/KEGG.txt",
  gene_expr = "example/slim_version/EXPRESSION.txt")
}
\arguments{
\item{gene_desc}{gene description}

\item{gene_kegg}{kegg to gene mapping}

\item{gene_expr}{expression mapping to gene}
}
\description{
This function allows compare expression of genes or protein abundancies to KEGG pathways
}
\examples{
cat_function()
}
\keyword{keggexpressionmapper}
